requirements.txt
flask>=2.0.0
firebase-admin>=6.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
pathlib>=1.0.1
datetime>=4.3